Rename package to br.com.concrete.requestmatcher
