+++ requestmatcher/src/main/AndroidManifest.xml
@@ -1,1 +1,1 @@
-<manifest package="br.com.concretesolutions.requestmatcher" />
+<manifest package="br.com.concrete.requestmatcher" />
