+++ requestmatcher/src/main/java/br/com/concrete/requestmatcher/exception/RequestAssertionException.java
@@ -1,0 +1,8 @@
+package br.com.concrete.requestmatcher.exception;
+
+public final class RequestAssertionException extends RuntimeException {
+
+    public RequestAssertionException(String message, Throwable e) {
+        super(message + " " + e.getMessage(), e);
+    }
+}
