--- requestmatcher/src/main/java/br/com/concretesolutions/requestmatcher/exception/RequestAssertionException.java
@@ -1,8 +1,0 @@
-package br.com.concretesolutions.requestmatcher.exception;
-
-public final class RequestAssertionException extends RuntimeException {
-
-    public RequestAssertionException(String message, Throwable e) {
-        super(message + " " + e.getMessage(), e);
-    }
-}
