+++ requestmatcher/src/test-common/java/br/com/concrete/requestmatcher/test/BaseTest.java
@@ -1,0 +1,9 @@
+package br.com.concrete.requestmatcher.test;
+
+
+import br.com.concrete.requestmatcher.RequestMatcherRule;
+
+public abstract class BaseTest {
+
+    protected abstract RequestMatcherRule getRequestMatcherRule();
+}
