--- requestmatcher/src/test-common/java/br/com/concretesolutions/requestmatcher/test/BaseTest.java
@@ -1,9 +1,0 @@
-package br.com.concretesolutions.requestmatcher.test;
-
-
-import br.com.concretesolutions.requestmatcher.RequestMatcherRule;
-
-public abstract class BaseTest {
-
-    protected abstract RequestMatcherRule getRequestMatcherRule();
-}
