+++ requestmatcher/src/test/java/br/com/concrete/requestmatcher/test/LocalExceptionWitouhtEnqueuesTest.java
@@ -1,0 +1,12 @@
+package br.com.concrete.requestmatcher.test;
+
+import br.com.concrete.requestmatcher.LocalTestRequestMatcherRule;
+import br.com.concrete.requestmatcher.RequestMatcherRule;
+
+public class LocalExceptionWitouhtEnqueuesTest extends ExceptionWitouhtEnqueuesTest {
+
+    @Override
+    protected RequestMatcherRule getRequestMatcherRule() {
+        return new LocalTestRequestMatcherRule();
+    }
+}
