--- requestmatcher/src/test/java/br/com/concretesolutions/requestmatcher/test/LocalExceptionWitouhtEnqueuesTest.java
@@ -1,12 +1,0 @@
-package br.com.concretesolutions.requestmatcher.test;
-
-import br.com.concretesolutions.requestmatcher.LocalTestRequestMatcherRule;
-import br.com.concretesolutions.requestmatcher.RequestMatcherRule;
-
-public class LocalExceptionWitouhtEnqueuesTest extends ExceptionWitouhtEnqueuesTest {
-
-    @Override
-    protected RequestMatcherRule getRequestMatcherRule() {
-        return new LocalTestRequestMatcherRule();
-    }
-}
