--- requestmatcher/src/test/java/br/com/concretesolutions/requestmatcher/test/LocalRequestMatcherRuleTest.java
@@ -1,12 +1,0 @@
-package br.com.concretesolutions.requestmatcher.test;
-
-import br.com.concretesolutions.requestmatcher.LocalTestRequestMatcherRule;
-import br.com.concretesolutions.requestmatcher.RequestMatcherRule;
-
-public class LocalRequestMatcherRuleTest extends RequestMatcherRuleTest {
-
-    @Override
-    protected RequestMatcherRule getRequestMatcherRule() {
-        return new LocalTestRequestMatcherRule();
-    }
-}
